:recycle: :: expiredAt 타입 Int -> Long으로 변경

+++ account/src/main/kotlin/com/project/indistraw/global/event/SaveRefreshTokenEvent.kt
@@ -5,5 +5,5 @@ import java.util.UUID
 data class SaveRefreshTokenEvent(
     val refreshToken: String,
     val accountIdx: UUID,
-    val expiredAt: Int
+    val expiredAt: Long
 )
